restructured package and changed final output to ModelHandle
